Cleanup that moves structure generation into a class
