+++ smithy-typescript-codegen/src/main/java/software/amazon/smithy/typescript/codegen/CodegenVisitor.java
@@ -31,7 +31,6 @@ import software.amazon.smithy.build.PluginContext;
 import software.amazon.smithy.codegen.core.Symbol;
 import software.amazon.smithy.codegen.core.SymbolDependency;
 import software.amazon.smithy.codegen.core.SymbolProvider;
-import software.amazon.smithy.codegen.core.SymbolReference;
 import software.amazon.smithy.model.Model;
 import software.amazon.smithy.model.knowledge.TopDownIndex;
 import software.amazon.smithy.model.neighbor.Walker;
@@ -44,7 +43,6 @@ import software.amazon.smithy.model.shapes.StringShape;
 import software.amazon.smithy.model.shapes.StructureShape;
 import software.amazon.smithy.model.shapes.UnionShape;
 import software.amazon.smithy.model.traits.EnumTrait;
-import software.amazon.smithy.model.traits.ErrorTrait;
 import software.amazon.smithy.model.traits.ProtocolsTrait;
 import software.amazon.smithy.typescript.codegen.integration.ProtocolGenerator;
 import software.amazon.smithy.typescript.codegen.integration.RuntimeClientPlugin;
@@ -186,141 +184,8 @@ class CodegenVisitor extends ShapeVisitor.Default<Void> {
      */
     @Override
     public Void structureShape(StructureShape shape) {
-        writers.useShapeWriter(shape, writer -> {
-            if (shape.hasTrait(ErrorTrait.class)) {
-                renderErrorStructure(shape, writer);
-            } else {
-                renderNonErrorStructure(shape, writer);
-            }
-        });
+        writers.useShapeWriter(shape, writer -> new StructureGenerator(model, symbolProvider, writer, shape).run());
         return null;
-    }
-
-    /**
-     * Renders a normal, non-error structure.
-     *
-     * <p>For example, given the following Smithy model:
-     *
-     * <pre>{@code
-     * namespace smithy.example
-     *
-     * structure Person {
-     *     @required
-     *     name: String,
-     *     age: Integer,
-     * }
-     * }</pre>
-     *
-     * <p>The following TypeScript is rendered:
-     *
-     * <pre>{@code
-     * import * as _smithy from "../lib/smithy";
-     *
-     * export interface Person {
-     *   __type?: "smithy.example#Person";
-     *   name: string | undefined;
-     *   age?: number | null;
-     * }
-     *
-     * export namespace Person {
-     *   export const ID = "smithy.example#Person";
-     *   export function isa(o: any): o is Person {
-     *     return _smithy.isa(o, ID);
-     *   }
-     * }
-     * }</pre>
-     *
-     * @param shape Structure to render.
-     * @param writer Writer to write to.
-     */
-    private void renderNonErrorStructure(StructureShape shape, TypeScriptWriter writer) {
-        Symbol symbol = symbolProvider.toSymbol(shape);
-        writer.writeShapeDocs(shape);
-
-        // Find symbol references with the "extends" property.
-        String extendsFrom = symbol.getReferences().stream()
-                .filter(ref -> ref.getProperty("extends").isPresent())
-                .map(SymbolReference::getAlias)
-                .collect(Collectors.joining(", "));
-
-        if (extendsFrom.isEmpty()) {
-            writer.openBlock("export interface $L {", symbol.getName());
-        } else {
-            writer.openBlock("export interface $L extends $L {", symbol.getName(), extendsFrom);
-        }
-
-        writer.write("__type?: $S;", shape.getId());
-        StructuredMemberWriter config = new StructuredMemberWriter(
-                model, symbolProvider, shape.getAllMembers().values());
-        config.writeMembers(writer, shape);
-        writer.closeBlock("}");
-        writer.write("");
-        renderStructureNamespace(shape, writer);
-    }
-
-    /**
-     * Error structures generate interfaces that extend from SmithyException
-     * and add the appropriate fault property.
-     *
-     * <p>Given the following Smithy structure:
-     *
-     * <pre>{@code
-     * namespace smithy.example
-     *
-     * @error("client")
-     * structure NoSuchResource {
-     *     @required
-     *     resourceType: String
-     * }
-     * }</pre>
-     *
-     * <p>The following TypeScript is generated:
-     *
-     * <pre>{@code
-     * import * as _smithy from "../lib/smithy";
-     *
-     * export interface NoSuchResource extends _smithy.SmithyException {
-     *   __type: "smithy.example#NoSuchResource";
-     *   $name: "NoSuchResource";
-     *   $fault: "client";
-     *   resourceType: string | undefined;
-     * }
-     *
-     * export namespace Person {
-     *   export const ID = "smithy.example#NoSuchResource";
-     *   export function isa(o: any): o is NoSuchResource {
-     *     return _smithy.isa(o, ID);
-     *   }
-     * }
-     * }</pre>
-     *
-     * @param shape Error shape being generated.
-     * @param writer Writer to write to.
-     */
-    private void renderErrorStructure(StructureShape shape, TypeScriptWriter writer) {
-        ErrorTrait errorTrait = shape.getTrait(ErrorTrait.class).orElseThrow(IllegalStateException::new);
-        Symbol symbol = symbolProvider.toSymbol(shape);
-        writer.writeShapeDocs(shape);
-        writer.openBlock("export interface $L extends _smithy.SmithyException {", symbol.getName());
-        writer.write("__type: $S;", shape.getId());
-        writer.write("$$name: $S;", shape.getId().getName());
-        writer.write("$$fault: $S;", errorTrait.getValue());
-        StructuredMemberWriter config = new StructuredMemberWriter(
-                model, symbolProvider, shape.getAllMembers().values());
-        config.writeMembers(writer, shape);
-        writer.closeBlock("}"); // interface
-        writer.write("");
-        renderStructureNamespace(shape, writer);
-    }
-
-    private void renderStructureNamespace(StructureShape shape, TypeScriptWriter writer) {
-        Symbol symbol = symbolProvider.toSymbol(shape);
-        writer.openBlock("export namespace $L {", "}", symbol.getName(), () -> {
-            writer.write("export const ID = $S;", shape.getId());
-            writer.openBlock("export function isa(o: any): o is $L {", "}", symbol.getName(), () -> {
-                writer.write("return _smithy.isa(o, ID);");
-            });
-        });
     }
 
     /**
